merge da dev para a main da segunda sprint

+++ frontend/src/pages/login/index.js
@@ -1,17 +1,26 @@
-import React, { useState } from "react";
+import React, { useState, useContext} from "react";
 import "./index.css";
 import axios from "axios";
 import logo from "../../assets/logo2.png";
 import { Link, useNavigate } from 'react-router-dom';
+import { AuthContext } from "../../contexts/authContext";
 import { ToastContainer, toast } from 'react-toastify';  
 import 'react-toastify/dist/ReactToastify.css';          
 
 function Login() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const { login } = useContext(AuthContext);
   const navigate = useNavigate();
   
-  localStorage.clear();
+  console.log(localStorage.getItem("historico"))
+  if (localStorage.getItem("historico")) {
+    localStorage.clear();
+  }
+   
+  if (localStorage.getItem("role")) {
+    localStorage.clear();
+  }
 
   const handleSubmit = async (event) => {
     event.preventDefault();
@@ -20,9 +29,6 @@ function Login() {
         email: email,
         password: password,
       });
-
-      console.log(loginResponse);
-      
 
       if (loginResponse.status === 200) {
         const token = loginResponse.data;
@@ -46,7 +52,7 @@ function Login() {
         localStorage.setItem('role', userLoged.role);
 
         if (userLoged.approved) {
-          navigate("/historico");
+          login(userLoged.role)
         } else {
           toast.warn("Usuário ainda não foi aprovado.");  
         }
